Patch pre-inlet velocity communication.
This replaces the `MPI_Isend` by `MPI_Send`, thereby replacing a
non-blocking send with a blocking send to ensure the velocity buffer
`vel` is not overwritten before the communication has taken place. The
send/receive loop inside `applyPreInletVelocityBoundary` is now only
using blocking send/receive operations, allowing to remove the
`MPI_Waitall` at the end of the routine.

The introduced change attempts to resolve an issue observed where the
velocity field of the pre-inlet was not properly communicated towards
the main domain, introducing spurious velocities at the inlet boundary.