Add googletest and update CMakeLists
This works towards adding `googletest` as testing framework and updating
the CMake files for compiling hemocell from its root.

The main `CMakeLists` is moved from `build/` to the root of the project,
enabling direct complication of the complete project. This includes
building the hemocell library, all examples, and all tests. The output
is then stored in `build/`. Note, to free up the build directory, the
Singularity definition is moved in its own directory.

To directly compile all examples requires a change to the `CMakeLists`
of the examples and the `example/` directory. As an initial working
example, the `example/pipeflow` is updated, and can now be compiled
directly from the root.

Compiling from the project root also provides the `googletest` framework
and compiles all tests located in `tests/`. Currently, only containing
two basic tests. These can be evaluated using `make test`, reporting on
the functionality of the code. Ultimately, these files could contain
sanity checks to be evaluated in CI/CD.