Move setup of dependencies to separate functions.
This moves the setup of MPI, HDF5, and `googletests` from the root
`CMakeLists.txt` towards separate functions included from `cmake/`. For
MPI and HDF5 default routines are used to discover and link these
dependencies. For `googletest` the function also takes care of
downloading the source code and compiling the test library.

Alongside the testing framework an additional target `make check` is
added, that invokes the tests and provides verbose output to support
investigation why tests might fail.

The `parmetis` and `metis` dependencies are removed (temporarily). There
is currently no test to validate if setup and linking is successful.
Once such example/test is added, the dependencies can be installed with
a separate function in `cmake/setup_xxx.cmake`.