Generalise build and compilation of examples.
Adds `examples/template` with a generic `CMakeLists.txt` that can be
used for all the example directories. This enables compilation from the
root directory of all examples, where the resulting executable is places
under the corresponding example directory.

The `clean.sh` and `compile.sh` are removed, as the compilation is done
through the root directory, which allows to clean the targets from
within `hemocell/build` directly.