Update cube benchmark for weak scaling.
The cubic domain is changed to a rectangular domain to better match a
constant atomic block size across different node allocations. The
current setup aims to maintain atomic blocks of `25^3` cells considering
nodes of 24 CPUs.

The corresponding configuration files and the `RBC.pos` are provided in
the compressed archive `config.tar.xz`. The RBC positioning provides a
grid-like layout of about ~15% up to domains of `1200x800x800` (2048
nodes of 24 CPU).