Compile library variant with interior viscosity.
When compiling all libraries, CMake will compile two variants of the
hemocell library: with and without the interior viscosity flag enabled.
Any example requiring interior viscosity will be linked with the
corresponding hemocell library to ensure the right functionality is
present. When compiling specific targets, only the required hemocell
library for the specific target is compiled.