Generalise XMF to X3D format conversion.
This generalises the XMF to X3D format conversion specialised the
onceCellShear to be applicable to any case. A small argument parser is
added to support passing multiple output directories to be processes.
Additionally, the view size and number of mesh smoothing iterations are
handled through command line arguments.

The script is executable and prints short usage information on
`./convert_xmf_to_x3d.py -h`.