Add documentation on visualisation and rendering.
A separate section is added to the Sphinx documentation
regarding visualisation of HemoCell's output.  This includes a brief
discussion on Paraview and a slightly more extensive discussion on
Blender rendering based on the `oneCellShear` example scripts. Also, an
additional reference to the Blender rendering is added to the
`oneCellShear` example.

The example commands in `scripts/visualization` are extended to a brief,
local documentation of the Blender rendering scripts.