Add support for running multiple validation tests.
To run multiple tests we are required to reinitialise the HemoCell
object multiple times when all tests are evaluated. This would result in
crashes for two reasons:

- Every initialisation would invoke `MPI_Init` and `MPI_Finalize`
- The HemoCell object did not unset its initialised flag, preventing new
  objects to be initialised after it was destroyed.

These issues are resolved by manually handling the MPI environment
before invoking all tests, i.e. invoke `plb::plbInit` manually in the
test runner `tests/main.cpp`. This ensures that a (global) MPI
environment is present that outlives the individual tests. Once all
tests are evaluated, this environment is then finalised. Doing so, we
ensure only a single call to `MPI_Init` and `MPI_Finalize` at the start
and end of the tests.

To ensure the HemoCell class does not attempt to call `plb::plbInit` on
its construction, an additional constructor was added that accepts the
new `HemoCell::MPIHandle` enum with either `MPI::Internal` or
`MPI::External`. The latter option indicates that the MPI environment is
handles externally, i.e. by the test driver, and `plb::plbInit` is not
called in that scenario.

Additionally, HemoCell's destructor ensures that
`global.hemoCellInitialize` is reset accordingly, such that new
instances might be constructed after its destruction.