Update stretchCell and include validation script.
The helper script `validation.sh` is added that allows for direct
validation of the implemented cell mechanics model. The script runs
through various evaluations of the `stretchCell` example and captures
the axial and transverse dimensions of the RBC over the iterations.
Afterwards, the axial and transverse force-displacements curves can be
recreated, illustrating the (validated) RBC cell mechanics.

The force-displacement curves can be compared with reference data
provided under the `validation/` directory, which correspond to the
validated force-displacement curves published at:
https://doi.org/10.3389/fphys.2017.00563.