Move function `rotateXYZ` from patch to HemoCell.
Originally, the patch introduces ``TriangularSurfaceMesh<T>::rotateXYZ``
directly on the ``TriangularSurfacMesh`` type in Palabos. To move this
function outside the patch, we introduce it as a separate function in
namespace `hemo` instead. Thus, a function as
`rotateTriangularMeshXYZ(TriangularSurfaceMesh<T>*, T, T, T)` is added
that applies the X, Y, Z rotation on a given triangular surface mesh.

This removes the code from the patch to be explicit in HemoCell.