Add `pos_to_vtk` tool for pos-file visualisation
This adds a Python-based tool for the direct visualisation of cell
`*.pos` files. The command-line utility provides means to visualise a
given `pos`-file for either red blood cells, platelets, or a combination
thereof. The visualisation is implemented using `PyVista` providing a
simple GUI window to explore the cell distribution.

Furthermore, when specifying an output file, the utility allows to
convert the `pos`-files towards XDMF format for further inspection in
Paraview.