Reduce messages send across pre-inlet boundary
This reduces the number of messages send that are required to transmit
cells from the periodic pre-inlet towards the main domain. Rather than
all atomic blocks sending their cells towards all other atomic blocks,
this implementation opts to only send messages when the atomic blocks
are overlapping. Thus, only neighbouring atomic blocks (neighbouring in
the sense that their bounding boxes overlap when projected in the
direction of the pre-inlet) send and receive messages.

This reduces the total number of messages that need to be send/received.
Additionally, due to less messages send, also time is saved in
(de)serialisation. This especially applies to ranks that did not
overlap, but that did receive messages containing cells, as these were
directly disregarded after extracting when realising they did not
overlap with the rank's bounding box.