Apply local buffer for communicating pre-inlet
This moves back to using the `MPI_Isend` rather than `MPI_Send` as
introduced in commit 4f4eab7c0c3ed8f5a7f406fa50bd17afce1ddbd3, as this
might cause deadlocks in certain configurations (software and/or
hardware).

To make sure `MPI_Isend` does not overwrite the local buffer, as was the
original problem resolved by the mentioned commit, an local vector is
allocated that temporarily stores the velocity values. This ensure the
buffer is not reduced in the loop and the right velocities are send to
the right target cells.